requirements.txt
python-dateutil~=2.8.1
PyJWT~=2.1.0
requests~=2.25.1
behave~=1.2.6
psycopg2~=2.8.6